Import IT++ 4.2

+++ itpp/comm/hammcode.h
@@ -5,24 +5,23 @@
  *
  * -------------------------------------------------------------------------
  *
- * IT++ - C++ library of mathematical, signal processing, speech processing,
- *        and communications classes and functions
+ * Copyright (C) 1995-2010  (see AUTHORS file for a list of contributors)
  *
- * Copyright (C) 1995-2009  (see AUTHORS file for a list of contributors)
+ * This file is part of IT++ - a C++ library of mathematical, signal
+ * processing, speech processing, and communications classes and functions.
  *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
+ * IT++ is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any
+ * later version.
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * IT++ is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
+ * You should have received a copy of the GNU General Public License along
+ * with IT++.  If not, see <http://www.gnu.org/licenses/>.
  *
  * -------------------------------------------------------------------------
  */
@@ -46,7 +45,7 @@ class Hamming_Code : public Channel_Code
 {
 public:
   //! Constructor for \c hamming(n,k). n = pow(2,m)-1 and k = pow(2,m)-m-1.
-  Hamming_Code(short m);
+  Hamming_Code(int m);
 
   //! Destructor
   virtual ~Hamming_Code() { }
@@ -69,15 +68,15 @@ public:
   virtual double get_rate() const { return static_cast<double>(k) / n; };
 
   //! Gets the code length \a n.
-  short get_n() const { return n; };
+  int get_n() const { return n; };
   //! Gets the number of information bits per code word, \a k.
-  short get_k() const { return k; };
+  int get_k() const { return k; };
   //! Gets the parity check matrix for the code.
   bmat get_H() const { return H; };
   //! Gets the generator matrix for the code.
   bmat get_G() const { return G; };
 private:
-  short n, k;
+  int n, k;
   bmat H, G;
   void generate_H(void);
   void generate_G(void);
